Refactor context, add react-icons and button props
